AI LEARNING COPILOT/backend/requirements.txt
# ================================
# CodeCore AI Backend - Python Dependencies
# ================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Requests (if needed for external APIs)
httpx==0.26.0

# Logging and Monitoring
python-json-logger==2.0.7
